[#21] Refactor: add index file to apis/user to import easaily
